requirements.txt
httpx>=0.27.0
supabase==2.9.1
pydantic-settings>=2.5.0
python-dotenv>=1.0.1
google-genai>=1.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
